requirements.txt
Django>=5.2.7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.3
drf-spectacular>=0.26
celery>=5.3
redis>=5.0
django-extensions>=3.2
whitenoise>=6.5
